Refactor temperature service; rename to AsyncAtmosphereService and enhance thread management for humidity, pressure, and storm calculations
Need to transform back into static

+++ src/main/java/net/Gabou/projectatmosphere/modules/pressure/forecast/PressureForecast.java
@@ -20,30 +20,30 @@ public class PressureForecast {
      * Scans a square (2*radius), finds one sample per biome,
      * then builds a 7-day pressure forecast using altitude, temp, humidity & storm spikes.
      */
-    public static Map<ResourceLocation, double[]> generateForecastAround(Level world,
+    public static Map<ResourceLocation, float[]> generateForecastAround(Level world,
                                                                          BlockPos center,
                                                                          int radius) {
         Set<ResourceLocation> found = new HashSet<>();
         Map<ResourceLocation, BlockPos> samples = new HashMap<>();
         AtmosphereUtils.findBiomes(world, center, radius, found, samples);
 
-        Map<ResourceLocation, double[]> forecast = new HashMap<>();
+        Map<ResourceLocation, float[]> forecast = new HashMap<>();
         for (var e : samples.entrySet()) {
             ResourceLocation biome = e.getKey();
             BlockPos pos = e.getValue();
-            double altitude = pos.getY();
-            double P0 = 1013.25 * Math.exp(-altitude / 8400.0);
+            float altitude = pos.getY();
+            float P0 = (float) (1013.25 * Math.exp(-altitude / 8400.0));
 
             float[][] tempWeek = TemperatureProfileManager.getWeeklyForecast(biome);
             if (tempWeek == null) continue;
 
-            double[] weekPress = new double[7];
+            float[] weekPress = new float[7];
             for (int d = 0; d < 7; d++) {
-                double Tmid = (tempWeek[d][0] + tempWeek[d][1]) / 2.0;
-                double deltaT = -0.5 * (Tmid - 15.0);            // ≈0.5 hPa per °C
-                double RH = HumidityManager.getAverageHumidity(biome, d);    // 0–100%
-                double deltaH = -0.05 * RH;                     // ≈0.05 hPa per %RH
-                double spike = StormManager.randomStormSpike(biome, d); // ±10 hPa
+                float Tmid = (float) ((tempWeek[d][0] + tempWeek[d][1]) / 2.0);
+                float deltaT = (float) (-0.5 * (Tmid - 15.0));            // ≈0.5 hPa per °C
+                float RH = HumidityManager.getAverageHumidity(biome, d);    // 0–100%
+                float deltaH = (float) (-0.05 * RH);                     // ≈0.05 hPa per %RH
+                float spike = (float) StormManager.randomStormSpike(biome, d); // ±10 hPa
                 weekPress[d] = P0 + deltaT + deltaH + spike;
             }
             forecast.put(biome, weekPress);
